For bigger numbers

+++ C++/CustomSolution.cpp
@@ -1,11 +1,11 @@
 #include <iostream>
 
 //a very fast solution that has a time complexity of O(sqrt(n))
-bool primalityTest(int num) {
+bool primalityTest(long long num) {
     if (num <= 1) return false;
     if (!(num%2)) return num==2;
 
-    for (int i = 3; i*i<=num; i+=2) {
+    for (long long i = 3; i*i<=num; i+=2) {
         if (!(num%i)) return false;
     }    
     return true;
@@ -13,7 +13,7 @@ bool primalityTest(int num) {
 
 
 int main() {
-    int num;
+    long long num;
     std::cout << "Enter a number to check if it's prime or not.\n";
     std::cin >> num;
     std::cout << "The number you entered is "<<(primalityTest(num) ? "prime" : "not prime")<< "\n";
